Move routes/ to pages/
