+++ app/client/router/routes.ts
@@ -1,8 +1,8 @@
 import { Router } from '@lit-labs/router';
 
+import * as about from '@/pages/about';
+import * as home from '@/pages/home';
 import { route } from '@/router/createRoute';
-import * as about from '@/routes/about';
-import * as home from '@/routes/home';
 
 export const routes = (router: Router) => [
   route('/', home.Page, { title: 'Project Template 2024' }),
@@ -12,7 +12,7 @@ export const routes = (router: Router) => [
     enter: async (params: { [key: string]: string | undefined }) => {
       const path = params[0];
       if (path === 'cats') {
-        const cats = await import('@/routes/cats');
+        const cats = await import('@/pages/cats');
         const { routes } = router;
         routes.splice(
           routes.length - 1,
